+++ docs/reference/changelog.rst
@@ -4,8 +4,14 @@
 Changelog
 *********
 
-5.2.0 (2025-MM-DD)
+5.2.0 (2025-04-25)
 ------------------
+
+Commands
+========
+
+- The ``test`` command now accepts paths to specific tests as well as the
+  ``--debug``, ``--shell`` and ``--shell-after`` parameters.
 
 Models
 ======
@@ -18,6 +24,8 @@ Services
 
 - The ``TestingService`` now sets environment variables containing the
   names of the generated artifact and resource files.
+
+For a complete list of commits, check out the `5.2.0`_ release on GitHub.
 
 5.1.0 (2025-04-24)
 ------------------
@@ -741,3 +749,4 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _5.0.3: https://github.com/canonical/craft-application/releases/tag/5.0.3
 .. _5.0.4: https://github.com/canonical/craft-application/releases/tag/5.0.4
 .. _5.1.0: https://github.com/canonical/craft-application/releases/tag/5.1.0
+.. _5.2.0: https://github.com/canonical/craft-application/releases/tag/5.2.0